feat - refactor team endpoints
